Reference_Implementation_Python/requirements.txt
numpy>=1.20.0
pycryptodome>=3.10.0
matplotlib